+++ extensions/datavirtualization/src/typings/ref.d.ts
@@ -3,7 +3,7 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-/// <reference path='../../../src/vscode-dts/vscode.d.ts'/>
-/// <reference path='../../../src/sql/azdata.d.ts'/>
-/// <reference path='../../../src/sql/azdata.proposed.d.ts'/>
+/// <reference path='../../../../src/vscode-dts/vscode.d.ts'/>
+/// <reference path='../../../../src/sql/azdata.d.ts'/>
+/// <reference path='../../../../src/sql/azdata.proposed.d.ts'/>
 /// <reference types='@types/node'/>
